requirements.txt
uvicorn[standard]
fastapi
sqlalchemy
bcrypt
python-jose[cryptography]
python-dotenv
PyMySQL
python-multipart
pytz
fastapi-utils
pydantic-settings